requirements.txt
SQLAlchemy==1.3.23
sqlalchemy-mixins==1.4
PyMySQL==1.0.2
alembic==1.6.2
pytest==6.2.1
pytest-factoryboy==2.0.3
